🏷️ fix types imports from a client component
